+++ CHANGES.rst
@@ -34,6 +34,7 @@ in progress
   and files on GitHub.
 - Tests: Add "roadrunner" tests, using a bunch of external resources. The tests
   will only check for successful invocation, and not verify the generated SQL.
+- Format: Add NetCDF input format
 
 
 2023-01-xx 0.0.0